pyproject.toml
[project]
name = "anbscouts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4.0",
    "faker>=37.1.0",
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "marshmallow-sqlalchemy>=1.4.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=5.2.1",
    "sqlalchemy>=2.0.40",
    "moviepy"
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
